+++ webpack.config.js
@@ -1,9 +1,9 @@
 const path = require("path");
 
 module.exports = {
-  entry: "./src/index.js",
+  entry: { main: "./src/index.js", media: "./media/editor.js" },
   output: {
-    filename: "main.js",
+    filename: "[name].js",
     path: path.resolve(__dirname, "dist"),
   },
   externals: {