+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-10-28
+
+Add two new settings, `sentence` and `characterList` for customizing the text shown in the preview.
+
+Text in the editable paragraph is remembered when tabbing away. Editor no longer reloads font every time you swap tabs.
+
+Add "Open Extension Settings" button to editor.
+
 ## [1.0.6] - 2023-03-29
 
 An option to disable ligatures.